refactor: move new state on toplevel
