+++ src/iokit/codec/zip.py
@@ -8,7 +8,7 @@ from typing import BinaryIO
 from zipfile import ZipFile
 
 from iokit.codec.base import Codec
-from iokit.utils.state_new import BufferedState, LoadedState, State
+from iokit.state_new import BufferedState, LoadedState, State
 from iokit.utils.time import Timestamp
